+++ forge16-defender/src/main/java/com/vecoo/legendcontrol_defender/LegendControlDefender.java
@@ -25,6 +25,8 @@ import net.minecraftforge.server.permission.PermissionAPI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.io.IOException;
+
 @Mod(LegendControlDefender.MOD_ID)
 public class LegendControlDefender {
     public static final String MOD_ID = "legendcontrol_defender";
@@ -64,7 +66,7 @@ public class LegendControlDefender {
         this.loadStorage();
 
         PermissionAPI.registerNode("minecraft.command.ltrust", DefaultPermissionLevel.OP, "/ltrust");
-        PermissionAPI.registerNode("minecraft.command.ltrust.reload", DefaultPermissionLevel.OP, "/ltrust reload");
+        PermissionAPI.registerNode("minecraft.command.ltrust.reload", DefaultPermissionLevel.OP, "/ltrust");
     }
 
     @SubscribeEvent
@@ -80,7 +82,7 @@ public class LegendControlDefender {
             this.locale = YamlConfigFactory.getInstance(LocaleConfig.class);
             this.storage = YamlConfigFactory.getInstance(StorageConfig.class);
             this.webhook = new DiscordWebhook(this.discord.getWebhookUrl());
-        } catch (Exception e) {
+        } catch (IOException e) {
             LOGGER.error("[LegendControl-Defender] Error load config.", e);
         }
     }