+++ forge16-defender/src/main/java/com/vecoo/legendcontrol_defender/storage/player/impl/PlayerDatabaseProvider.java
@@ -36,7 +36,6 @@ public class PlayerDatabaseProvider implements PlayerProvider {
         return this.map.get(playerUUID);
     }
 
-
     @Override
     public void updatePlayerStorage(PlayerStorage storage) {
         this.map.put(storage.getUUID(), storage);