+++ forge16/src/main/java/com/vecoo/legendcontrol/LegendControl.java
@@ -29,6 +29,8 @@ import net.minecraftforge.server.permission.PermissionAPI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.io.IOException;
+
 @Mod(LegendControl.MOD_ID)
 public class LegendControl {
     public static final String MOD_ID = "legendcontrol";
@@ -73,7 +75,7 @@ public class LegendControl {
         this.loadStorage();
 
         PermissionAPI.registerNode("minecraft.command.checkleg", DefaultPermissionLevel.OP, "/checkleg");
-        PermissionAPI.registerNode("minecraft.command.checkleg.modify", DefaultPermissionLevel.OP, "Modify version /checkleg");
+        PermissionAPI.registerNode("minecraft.command.checkleg.modify", DefaultPermissionLevel.OP, "/checkleg");
         PermissionAPI.registerNode("minecraft.command.lc", DefaultPermissionLevel.OP, "/lc");
     }
 
@@ -91,7 +93,7 @@ public class LegendControl {
             this.storage = YamlConfigFactory.getInstance(StorageConfig.class);
             this.discord = YamlConfigFactory.getInstance(DiscordConfig.class);
             this.webhook = new DiscordWebhook(this.discord.getWebhookUrl());
-        } catch (Exception e) {
+        } catch (IOException e) {
             LOGGER.error("[LegendControl] Error load config.", e);
         }
     }