+++ neoforge21-defender/src/main/java/com/vecoo/legendcontrol_defender/api/factory/LegendControlFactory.java
@@ -1,9 +1,7 @@
 package com.vecoo.legendcontrol_defender.api.factory;
 
 import com.vecoo.legendcontrol_defender.LegendControlDefender;
-import com.vecoo.legendcontrol_defender.api.events.PlayerTrustEvent;
 import com.vecoo.legendcontrol_defender.storage.player.PlayerStorage;
-import net.neoforged.neoforge.common.NeoForge;
 
 import java.util.Map;
 import java.util.Set;
@@ -23,22 +21,16 @@ public class LegendControlFactory {
             return LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).getPlayersTrust().contains(targetUUID);
         }
 
-        public static void addPlayerTrust(UUID playerUUID, UUID targetUUID, boolean update) {
-            if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Add(playerUUID, targetUUID)).isCanceled()) {
-                LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).addPlayerTrust(targetUUID, update);
-            }
+        public static void addPlayerTrust(UUID playerUUID, UUID targetUUID) {
+            LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).addPlayerTrust(targetUUID, true);
         }
 
-        public static void removePlayerTrust(UUID playerUUID, UUID targetUUID, boolean update) {
-            if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Remove(playerUUID, targetUUID)).isCanceled()) {
-                LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).removePlayerTrust(targetUUID, update);
-            }
+        public static void removePlayerTrust(UUID playerUUID, UUID targetUUID) {
+            LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).removePlayerTrust(targetUUID, true);
         }
 
-        public static void removePlayersTrust(UUID playerUUID, boolean update) {
-            if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Clear(playerUUID)).isCanceled()) {
-                LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).clearPlayersTrust(update);
-            }
+        public static void removePlayersTrust(UUID playerUUID) {
+            LegendControlDefender.getInstance().getPlayerProvider().getPlayerStorage(playerUUID).clearPlayersTrust(true);
         }
     }
 }