+++ neoforge21-defender/src/main/java/com/vecoo/legendcontrol_defender/storage/player/PlayerStorage.java
@@ -1,6 +1,8 @@
 package com.vecoo.legendcontrol_defender.storage.player;
 
 import com.vecoo.legendcontrol_defender.LegendControlDefender;
+import com.vecoo.legendcontrol_defender.api.events.PlayerTrustEvent;
+import net.neoforged.neoforge.common.NeoForge;
 
 import java.util.Set;
 import java.util.UUID;
@@ -24,26 +26,32 @@ public class PlayerStorage {
     }
 
     public void addPlayerTrust(UUID playerUUID, boolean update) {
-        this.playersTrust.add(playerUUID);
+        if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Add(this.uuid, playerUUID)).isCanceled()) {
+            this.playersTrust.add(playerUUID);
 
-        if (update) {
-            LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            if (update) {
+                LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            }
         }
     }
 
     public void removePlayerTrust(UUID playerUUID, boolean update) {
-        this.playersTrust.remove(playerUUID);
+        if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Remove(this.uuid, playerUUID)).isCanceled()) {
+            this.playersTrust.remove(playerUUID);
 
-        if (update) {
-            LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            if (update) {
+                LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            }
         }
     }
 
     public void clearPlayersTrust(boolean update) {
-        this.playersTrust.clear();
+        if (!NeoForge.EVENT_BUS.post(new PlayerTrustEvent.Clear(this.uuid)).isCanceled()) {
+            this.playersTrust.clear();
 
-        if (update) {
-            LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            if (update) {
+                LegendControlDefender.getInstance().getPlayerProvider().updatePlayerStorage(this);
+            }
         }
     }
 }