+++ neoforge21-defender/src/main/java/com/vecoo/legendcontrol_defender/util/WebhookUtils.java
@@ -13,12 +13,16 @@ public class WebhookUtils {
         if (!discordConfig.getWebhookUrl().isEmpty()) {
             try {
                 LegendControlDefender.getInstance().getWebhook().sendEmbed(discordConfig.getWebhookTitleExpiredDefender()
-                                .replace("%shiny%", pixelmonEntity.getPokemon().isShiny() ? ":star2: " : ""),
+                                .replace("%shiny%", getShinyText(pixelmonEntity)),
                         discordConfig.getWebhookDescriptionExpiredDefender()
                                 .replace("%pokemon%", pixelmonEntity.getPokemonName()), Utils.pokemonImage(pixelmonEntity), discordConfig.getWebhookColor());
             } catch (IOException e) {
                 LegendControlDefender.getLogger().error("[LegendControl-Defender] Error send Discord webhook", e);
             }
         }
+    }
+
+    private static String getShinyText(PixelmonEntity pixelmonEntity) {
+        return pixelmonEntity.getPokemon().isShiny() ? ":star2: " : "";
     }
 }