+++ neoforge21/src/main/java/com/vecoo/legendcontrol/api/events/ChangeChanceLegendEvent.java
@@ -3,11 +3,11 @@ package com.vecoo.legendcontrol.api.events;
 import net.neoforged.bus.api.Event;
 import net.neoforged.bus.api.ICancellableEvent;
 
-public class ChanceLegendEvent extends Event implements ICancellableEvent {
+public class ChangeChanceLegendEvent extends Event implements ICancellableEvent {
     private final String source;
-    private final float chance;
+    private float chance;
 
-    public ChanceLegendEvent(String source, float chance) {
+    public ChangeChanceLegendEvent(String source, float chance) {
         this.source = source;
         this.chance = chance;
     }
@@ -18,5 +18,9 @@ public class ChanceLegendEvent extends Event implements ICancellableEvent {
 
     public float getChance() {
         return this.chance;
+    }
+
+    public void setChance(float chance) {
+        this.chance = chance;
     }
 }