+++ changelog.md
@@ -1,1 +1,3 @@
-* Change license.
+* Added blacklist player.
+* Improving API.
+* Other minor changes and improvements.